PR review: rename to erc20, split into balance/metadata, use Button.Symbol
